refactor(di): moves `di` to separate `package`
