+++ cli/pull_command.go
@@ -11,8 +11,8 @@ package main
 
 import (
 	"fmt"
+	"github.com/Fitbit/smartling/di"
 	"github.com/Fitbit/smartling/model"
-	"github.com/Fitbit/smartling/service"
 	"github.com/fatih/color"
 	"gopkg.in/go-playground/pool.v3"
 	"gopkg.in/urfave/cli.v1"
@@ -40,7 +40,7 @@ var pullCommand = cli.Command{
 	Before: func(c *cli.Context) error {
 		return invokeActions([]action{
 			ensureMetadataAction,
-			injectContainerAction,
+			injectDiContainerAction,
 			injectProjectConfigAction,
 			validateProjectConfigAction,
 			injectAuthTokenAction,
@@ -55,7 +55,7 @@ var pullCommand = cli.Command{
 
 		batch := p.Batch()
 
-		container := c.App.Metadata[containerKey].(*service.Container)
+		container := c.App.Metadata[containerKey].(*di.Container)
 		authToken := c.App.Metadata[authTokenKey].(*model.AuthToken)
 		projectConfig := c.App.Metadata[projectConfigKey].(*model.ProjectConfig)
 		retrievalType := c.String("retrieval-type")
