+++ di/setup.go
@@ -1,4 +1,4 @@
-// Copyright 2016, Fitbit, Inc.
+// Copyright 2017, Fitbit, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -7,43 +7,40 @@
 //
 // Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and limitations under the License.
-package service
+package di
 
 import (
 	"github.com/Fitbit/smartling/repository"
 	"github.com/Fitbit/smartling/rest"
+	"github.com/Fitbit/smartling/service"
 	"github.com/facebookgo/inject"
 )
 
-type Container struct {
-	ProjectConfigService ProjectConfigService `inject:"DefaultProjectConfigService"`
-	AuthService          AuthService          `inject:"DefaultAuthService"`
-	FileService          FileService          `inject:"DefaultFileService"`
-}
-
-func (c *Container) SetUp(filename string) error {
+func Setup(opts *Options) (*Container, error) {
 	var g inject.Graph
+
+	c := Container{}
 
 	err := g.Provide(
 		&inject.Object{
-			Value: c,
+			Value: &c,
 		},
 		&inject.Object{
 			Value: &repository.YmlProjectConfigRepository{
-				Filename: filename,
+				Filename: opts.Filename,
 			},
 			Name: "YmlProjectConfigRepository",
 		},
 		&inject.Object{
-			Value: &DefaultProjectConfigService{},
+			Value: &service.DefaultProjectConfigService{},
 			Name:  "DefaultProjectConfigService",
 		},
 		&inject.Object{
-			Value: &DefaultAuthService{},
+			Value: &service.DefaultAuthService{},
 			Name:  "DefaultAuthService",
 		},
 		&inject.Object{
-			Value: &DefaultFileService{},
+			Value: &service.DefaultFileService{},
 			Name:  "DefaultFileService",
 		},
 		&inject.Object{
@@ -56,5 +53,5 @@ func (c *Container) SetUp(filename string) error {
 		err = g.Populate()
 	}
 
-	return err
+	return &c, err
 }
